Refactor ListingModal into separate components
